Move cqsh out of components
